Add Laravel NOVA
This commit adds new font files under the 'public/vendor/nova/fonts' directory. Simultaneously, it removes the CityController, CountryController, StateController, and CityPolicy classes, as well as the UpdateDepartmentRequest class. Additionally, it introduces a migration for creating an industries table and adds an app.css file in the 'public/vendor/nova' directory.